+++ src/shared/utils/cpfValidator.ts
@@ -1,0 +1,19 @@
+export function validateCPF(cpf: string): boolean {
+  cpf = cpf.replace(/[^\d]+/g, '');
+
+  if (cpf.length !== 11 || !!cpf.match(/(\d)\1{10}/)) {
+    return false;
+  }
+
+  const cpfsDigits = cpf.split('').map((el) => +el);
+
+  const rest = (count: number) =>
+    ((cpfsDigits
+      .slice(0, count - 12)
+      .reduce((soma, el, index) => soma + el * (count - index), 0) *
+      10) %
+      11) %
+    10;
+
+  return rest(10) === cpfsDigits[9] && rest(11) === cpfsDigits[10];
+}